spin out tides, magnetic braking, grav rad, winds into their own files
